azure_function/requirements.txt
# Azure Functions Python requirements
azure-functions
azure-storage-blob